--- .env
@@ -1,1 +1,0 @@
-REACT_APP_API_URI="http://localhost:3000"
